requirements.txt
python-gitlab==2.1.2
prometheus_client==0.7.1
pyyaml==5.3